refactor: 일급 컬렉션(First Class Collection) 적용 및 폴더 구조 개선
