+++ src/main/java/com/burntoburn/easyshift/repository/schedule/ShiftTemplateRepository.java
@@ -1,6 +1,6 @@
 package com.burntoburn.easyshift.repository.schedule;
 
-import com.burntoburn.easyshift.entity.schedule.ShiftTemplate;
+import com.burntoburn.easyshift.entity.templates.ShiftTemplate;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
