+++ src/main/java/com/burntoburn/easyshift/service/shift/ShiftTemplateService.java
@@ -1,6 +1,6 @@
 package com.burntoburn.easyshift.service.shift;
 
-import com.burntoburn.easyshift.entity.schedule.ShiftTemplate;
+import com.burntoburn.easyshift.entity.templates.ShiftTemplate;
 import java.util.List;
 
 public interface ShiftTemplateService {
