+++ src/test/java/com/burntoburn/easyshift/service/shift/imp/ShiftTemplateServiceImpTest.java
@@ -1,6 +1,6 @@
 package com.burntoburn.easyshift.service.shift.imp;
 
-import com.burntoburn.easyshift.entity.schedule.ShiftTemplate;
+import com.burntoburn.easyshift.entity.templates.ShiftTemplate;
 import com.burntoburn.easyshift.repository.schedule.ShiftTemplateRepository;
 import com.burntoburn.easyshift.service.shift.ShiftTemplateService;
 import org.junit.jupiter.api.DisplayName;
